kimhs 11.25. 주기 커밋
