+++ src/components/test/UseStateTest.js
@@ -15,7 +15,7 @@ function LoginForm() {
     navigate('/board', { state: { testId, testPw }});
   }
   return (
-    <div id='login' className='wrap'>
+    <div id='login' className='Wrap'>
       <div id='loginArea'>
           <form>
             <div className='inputFields idField'>
@@ -23,17 +23,17 @@ function LoginForm() {
               <Form.Control defaultValue={testId} type='id' id='inputId' placeholder='User ID' onChange={(e) => setTestId(e.target.value)}/>
             </div>
             <div className='inputFields pwField'>
-              <Form.Label htmlFor='inputPw'  >Password</Form.Label>
+              <Form.Label htmlFor="inputPw"  >Password</Form.Label>
               <Form.Control defaultValue={testPw}
-              type='password'
-              id='inputPw' placeholder='Password' onChange={(e) => setTestPw(e.target.value)}
-              aria-describedby='passwordHelpBlock'
+              type="password"
+              id="inputPw" placeholder='Password' onChange={(e) => setTestPw(e.target.value)}
+              aria-describedby="passwordHelpBlock"
               />
-              <Form.Text id='passwordHelpBlock' muted>
+              <Form.Text id="passwordHelpBlock" muted>
               useState로 컴포넌트끼리 변수 넘겨 주는 Test Page
               </Form.Text>
             </div>
-            <Button type='submit' variant='primary' onClick={handleSubmit}>확인</Button>
+            <Button type='submit' variant="primary" onClick={handleSubmit}>확인</Button>
           </form>
           {/* react에서는 input-label을 이을 때 label id 값이 아닌 htmlFor를 사용한다. */}
       </div>
