--- src/utils/datas/test.json
@@ -1,88 +1,0 @@
-
-[[
-  {
-    "본부": "델사업본부",
-    "팀": "영업1팀",
-    "담당자": "1",
-    "매출": "20,000",
-    "매출이익": "20,000"
-  },
-  {
-    "본부": "델사업본부",
-    "팀": "영업1팀",
-    "담당자": "2",
-    "매출": "20,000",
-    "매출이익": "20,000"
-  },
-  {
-    "본부": "델사업본부",
-    "팀": "영업1팀",
-    "담당자": "3",
-    "매출": "20,000",
-    "매출이익": "20,000"
-  },
-  {
-    "본부": "델사업본부",
-    "팀": "영업2팀",
-    "담당자": "4",
-    "매출": "20,000",
-    "매출이익": "20,000"
-  },
-  {
-    "본부": "델사업본부",
-    "팀": "영업3팀",
-    "담당자": "5",
-    "매출": "20,000",
-    "매출이익": "20,000"
-  },
-  {
-    "본부": "델사업본부",
-    "팀": "영업3팀",
-    "담당자": "6",
-    "매출": "20,000",
-    "매출이익": "20,000"
-  },
-  {
-    "본부": "델사업본부",
-    "팀": "영업4팀",
-    "담당자": "7",
-    "매출": "20,000",
-    "매출이익": "20,000"
-  },
-  {
-    "본부": "델사업본부",
-    "팀": "영업4팀",
-    "담당자": "8",
-    "매출": "20,000",
-    "매출이익": "20,000"
-  },
-  {
-    "본부": "델사업본부",
-    "팀": "영업4팀",
-    "담당자": "9",
-    "매출": "20,000",
-    "매출이익": "20,000"
-  }
-],
-[
-  {
-    "Header": "본부",
-    "accessor": "본부"
-  },
-  {
-    "Header": "팀",
-    "accessor": "팀"
-  },
-  {
-    "Header": "담당자",
-    "accessor": "담당자"
-  },
-  {
-    "Header": "매출",
-    "accessor": "매출"
-  },
-  {
-    "Header": "매출이익",
-    "accessor": "매출이익"
-  }
-]]
